tradesbot/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Twilio
twilio==9.4.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Redis
redis[hiredis]==5.2.1

# MinIO (S3)
minio==7.2.12

# Groq (ASR + LLM)
groq==0.15.0
httpx==0.28.1

# LangChain + LLM
langchain==0.3.14
langchain-groq==0.2.4
langchain-community==0.3.14

# Vector DB
chromadb==0.5.23

# PDF Generation
fpdf2==2.8.3

# Audio processing
pydub==0.25.1

# Utilities
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
structlog==24.4.0
tenacity==9.0.0
